+++ swing/common-model/src/test/java/is/codion/swing/common/model/component/list/FilterListSelectionTest.java
@@ -57,8 +57,8 @@ public class FilterListSelectionTest {
 											}
 										})
 										.items(() -> data)
+										.refresh(true)
 										.build();
-		tableModel.items().refresh();
 
 		testModel = filterListSelection(tableModel.items().included());
 	}